change FLOAT to DECIMAL for price, STRING to TEXT for imageUrl

+++ server/db/models/historial.js
@@ -6,7 +6,7 @@ const Historial = db.define('historial', {
     type: Sequelize.INTEGER
   },
   price: {
-    type: Sequelize.FLOAT
+    type: Sequelize.DECIMAL(5, 2)
   }
 })
